+++ src/main/java/io/quarkus/sample/agent/AgentMessage.java
@@ -1,0 +1,5 @@
+package io.quarkus.sample.agent;
+
+public record AgentMessage(Kind kind, String todoId, String payload) {
+    
+}
