+++ src/main/java/io/quarkus/sample/agent/Kind.java
@@ -1,0 +1,5 @@
+package io.quarkus.sample.agent;
+
+public enum Kind {
+    cancel, activity_log, agent_request, user_message
+}
